requirements.txt
supabase==2.3.4
moviepy==1.0.3
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0
validators==0.22.0 